SCT classes factory implemented
The SctFactory builds and provides the Payment, Transaction and IdHolder classes
