+++ sepacbi/__init__.py
@@ -7,7 +7,7 @@ Import classes exposed as the API to consumer modules.
 __copyright__ = 'Copyright (c) 2014 Emanuele Pucciarelli, C.O.R.P. s.n.c.'
 __license__ = '3-clause BSD'
 
-from .factory import SctFactory
+from .sct_factory import SctFactory
 # from .entity import IdHolder
 # from .payment import Payment
 from .rmtinfo import Document, Invoice, CreditNote, DebitNote, Text
