+++ sepacbi/entity.py
@@ -23,7 +23,7 @@ class AddressFormatError(Exception):
     Raised when an Address is created with unsuitable arguments.
     """
 
-class SctAddress(AttributeCarrier):
+class Address(AttributeCarrier):
     """
     The postal address of an entity for the SEPA SCT request mode.
     """
@@ -64,77 +64,77 @@ def emit_id_tag(id_code, id_type=None):
     return othr
 
 
-class SctIdHolder(AttributeCarrier):
+class IdHolder(AttributeCarrier):
     # pylint: disable=no-member
     """
     Describes an Initiator.
     Must have a CUC; may have an Italian tax code.
     """
-    allowed_args = ('name', 'cf', 'code', 'private', 'cuc', 'address',
-                    'country', 'sia_code')
+    # allowed_args = ('name', 'cf', 'code', 'private', 'cuc', 'address',
+    #                 'country', 'sia_code')
 
     def __init__(self, **kwargs):
         self.private = False
-        super(SctIdHolder, self).__init__(**kwargs)
+        super(IdHolder, self).__init__(**kwargs)
 
-    def perform_checks(self):
-        # pylint: disable=access-member-before-definition
-        # pylint: disable=attribute-defined-outside-init
-        "Check argument lengths."
-        if hasattr(self, 'name'):
-            self.max_length('name', 70)
-        if hasattr(self, 'cf'):
-            assert not hasattr(self, 'code')
-            self.max_length('cf', 16)
-        if hasattr(self, 'cuc'):
-            self.max_length('cuc', 35)
-        if hasattr(self, 'address'):
-            if isinstance(self.address, (list, tuple)):
-                self.address = SctAddress(*self.address)
-            assert isinstance(self.address, SctAddress)
-        if hasattr(self, 'country'):
-            self.length('country', 2)
+    # def perform_checks(self):
+    #     # pylint: disable=access-member-before-definition
+    #     # pylint: disable=attribute-defined-outside-init
+    #     "Check argument lengths."
+    #     if hasattr(self, 'name'):
+    #         self.max_length('name', 70)
+    #     if hasattr(self, 'cf'):
+    #         assert not hasattr(self, 'code')
+    #         self.max_length('cf', 16)
+    #     if hasattr(self, 'cuc'):
+    #         self.max_length('cuc', 35)
+    #     if hasattr(self, 'address'):
+    #         if isinstance(self.address, (list, tuple)):
+    #             self.address = Address(*self.address)
+    #         assert isinstance(self.address, Address)
+    #     if hasattr(self, 'country'):
+    #         self.length('country', 2)
 
-    def emit_tag(self, tag=None, as_initiator=False):
-        """
-        Emit a subtree for an entity, using the supplied tag for the root
-        element. If the identity is the Initiator's, emit the CUC as well.
-        """
-        if as_initiator:
-            tag = 'InitgPty'
-        root = etree.Element(tag)
-
-        # Name
-        if hasattr(self, 'name'):
-            name = etree.SubElement(root, 'Nm')
-            name.text = self.name
-
-        # Address
-        if hasattr(self, 'address') and not as_initiator:
-            root.append(self.address.__tag__())
-
-        # ID
-        idtag = etree.SubElement(root, 'Id')
-        if self.private:
-            id_container = 'PrvtId'
-        else:
-            id_container = 'OrgId'
-        orgid = etree.SubElement(idtag, id_container)
-
-        # CUC
-        if as_initiator:
-            if not hasattr(self, 'cuc'):
-                raise MissingCUCError
-            orgid.append(emit_id_tag(self.cuc, 'CBI'))
-
-        # Tax code
-        if hasattr(self, 'cf'):
-            orgid.append(emit_id_tag(self.cf, 'ADE'))
-
-        if hasattr(self, 'code'):
-            orgid.append(emit_id_tag(self.code, None))
-
-        if not as_initiator and hasattr(self, 'country'):
-            etree.SubElement(root, 'CtryOfRes').text = self.country
-
-        return root
+    # def emit_tag(self, tag=None, as_initiator=False):
+    #     """
+    #     Emit a subtree for an entity, using the supplied tag for the root
+    #     element. If the identity is the Initiator's, emit the CUC as well.
+    #     """
+    #     if as_initiator:
+    #         tag = 'InitgPty'
+    #     root = etree.Element(tag)
+    #
+    #     # Name
+    #     if hasattr(self, 'name'):
+    #         name = etree.SubElement(root, 'Nm')
+    #         name.text = self.name
+    #
+    #     # Address
+    #     if hasattr(self, 'address') and not as_initiator:
+    #         root.append(self.address.__tag__())
+    #
+    #     # ID
+    #     idtag = etree.SubElement(root, 'Id')
+    #     if self.private:
+    #         id_container = 'PrvtId'
+    #     else:
+    #         id_container = 'OrgId'
+    #     orgid = etree.SubElement(idtag, id_container)
+    #
+    #     # CUC
+    #     if as_initiator:
+    #         if not hasattr(self, 'cuc'):
+    #             raise MissingCUCError
+    #         orgid.append(emit_id_tag(self.cuc, 'CBI'))
+    #
+    #     # Tax code
+    #     if hasattr(self, 'cf'):
+    #         orgid.append(emit_id_tag(self.cf, 'ADE'))
+    #
+    #     if hasattr(self, 'code'):
+    #         orgid.append(emit_id_tag(self.code, None))
+    #
+    #     if not as_initiator and hasattr(self, 'country'):
+    #         etree.SubElement(root, 'CtryOfRes').text = self.country
+    #
+    #     return root
